# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/portfolio_join_image_files": {
            "get": {
                "tags": [
                    "Portfolio_Join_Image_File"
                ],
                "summary": "all joinfile",
                "description": "portfolio_join_image_files",
                "operationId": "index_file",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "image_file_id": {
                                            "type": "string",
                                            "example": "id"
                                        },
                                        "portfolio_id": {
                                            "type": "string",
                                            "example": "id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Portfolio_Join_Image_File"
                ],
                "summary": "portfolio join image file a new data",
                "description": "new join image file",
                "operationId": "6d57bcf99162a28507f7e45a1be592d8",
                "requestBody": {
                    "description": "POrtfolio_Join_Image_File",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image_file_id",
                                    "portfolio_id"
                                ],
                                "properties": {
                                    "image_file_id": {
                                        "type": "string",
                                        "example": "id"
                                    },
                                    "portfolio_id": {
                                        "type": "string",
                                        "example": "id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/portfolio_join_image_files/{portfolio_join_image_file}": {
            "get": {
                "tags": [
                    "Portfolio_Join_Image_File"
                ],
                "summary": "join image file",
                "description": "Returns file data",
                "operationId": "4baa3411dc15ab52c2671edc9ed683f0",
                "parameters": [
                    {
                        "name": "portfolio_join_image_file",
                        "in": "path",
                        "description": "file id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Portfolio_Join_Image_File"
                ],
                "summary": "image_file a new data",
                "description": "Portfolio_Join_Image_File_crud",
                "operationId": "1949bba76ea4f5f80522ab60c0c3eb4c",
                "parameters": [
                    {
                        "name": "portfolio_join_image_file_id",
                        "in": "path",
                        "description": "Id of file",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "portfolio join image file",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image_file_id",
                                    "portfolio_id"
                                ],
                                "properties": {
                                    "image_file_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "portfolio_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Portfolio_Join_Image_File"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "41f56c5559c4545dcfb7e64a67ffc493",
                "parameters": [
                    {
                        "name": "portfolio_join_image_file",
                        "in": "path",
                        "description": "File id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}